network instances used in the paper
The file contains networks in four lines.

The first line contains five values: NetID  #ofNodes  #ofDirectedEdges Tx Rx

The second line lists the intial energy of all nodes.

The third line lists all edges. The graph is undirected, and each edge is listed in ascending order of node ID. For example, edge {3, 2} is written as 2 3

The fourth line lists the 2D coordinates of all  nodes. This is helpfull to view the network in 2D plane.